refactor/shared_library_support    - this is commit refactors the include folder and moves includes from src/ folder to include/    - add support for shared libraries    - isolate definitions in cmake to one project(no more global add_definition and enable_language)    - changed CamelCase to snake_case classes in file and network includes    - added open and open_throws static functions to file_handle    - removed bad constructor for file_handle    - make async_filebuf and async_iofstream forcefully static in any build    - removed custom new and delete definitions for correct shared library work, now welcome tagged new and delete in the include/allocator.hpp    - add allocate and free functions to include/allocator.hpp    - add custom allocator struct to  include/allocator.hpp    - make the experimental preemptive scheduler disabled by default    - add spin_lock to include/threading.hpp

+++ src/tasks/classes/task/scheduler.cpp
@@ -4,7 +4,7 @@
 // (See accompanying file LICENSE or copy at
 // http://www.boost.org/LICENSE_1_0.txt)
 
-#include <tasks.hpp>
+#include <task.hpp>
 #include <tasks/_internal.hpp>
 
 namespace fast_task::scheduler {
@@ -134,7 +134,7 @@ namespace fast_task::scheduler {
             glob.binded_workers.erase(id);
         }
         for (std::shared_ptr<task>& task : transfer_tasks) {
-            get_data(task).bind_to_worker_id = -1;
+            get_data(task).bind_to_worker_id = (uint16_t)-1;
             transfer_task(task);
         }
     }
